merge new version

+++ activemq-broker/src/main/java/org/apache/activemq/util/LockFile.java
@@ -16,6 +16,9 @@
  */
 package org.apache.activemq.util;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import java.io.File;
 import java.io.IOException;
 import java.io.RandomAccessFile;
@@ -32,11 +35,16 @@ public class LockFile {
 
     private static final boolean DISABLE_FILE_LOCK = Boolean.getBoolean("java.nio.channels.FileLock.broken");
     final private File file;
+    private long lastModified;
 
     private FileLock lock;
-    private RandomAccessFile readFile;
+    private RandomAccessFile randomAccessLockFile;
     private int lockCounter;
     private final boolean deleteOnUnlock;
+    private volatile boolean locked;
+    private String lockSystemPropertyName = "";
+
+    private static final Logger LOG = LoggerFactory.getLogger(LockFile.class);
 
     public LockFile(File file, boolean deleteOnUnlock) {
         this.file = file;
@@ -57,25 +65,31 @@ public class LockFile {
 
         IOHelper.mkdirs(file.getParentFile());
         synchronized (LockFile.class) {
-            if (System.getProperty(getVmLockKey()) != null) {
-                throw new IOException("File '" + file + "' could not be locked as lock is already held for this jvm.");
+            lockSystemPropertyName = getVmLockKey();
+            if (System.getProperty(lockSystemPropertyName) != null) {
+                throw new IOException("File '" + file + "' could not be locked as lock is already held for this jvm. Value: " + System.getProperty(lockSystemPropertyName));
             }
-            System.setProperty(getVmLockKey(), new Date().toString());
+            System.setProperty(lockSystemPropertyName, new Date().toString());
         }
         try {
             if (lock == null) {
-                readFile = new RandomAccessFile(file, "rw");
+                randomAccessLockFile = new RandomAccessFile(file, "rw");
                 IOException reason = null;
                 try {
-                    lock = readFile.getChannel().tryLock(0, Math.max(1, readFile.getChannel().size()), false);
+                    lock = randomAccessLockFile.getChannel().tryLock(0, Math.max(1, randomAccessLockFile.getChannel().size()), false);
                 } catch (OverlappingFileLockException e) {
                     reason = IOExceptionSupport.create("File '" + file + "' could not be locked.", e);
                 } catch (IOException ioe) {
                     reason = ioe;
                 }
                 if (lock != null) {
+                    //track lastModified only if we are able to successfully obtain the lock.
+                    randomAccessLockFile.writeLong(System.currentTimeMillis());
+                    randomAccessLockFile.getChannel().force(true);
+                    lastModified = file.lastModified();
                     lockCounter++;
-                    System.setProperty(getVmLockKey(), new Date().toString());
+                    System.setProperty(lockSystemPropertyName, new Date().toString());
+                    locked = true;
                 } else {
                     // new read file for next attempt
                     closeReadFile();
@@ -89,7 +103,7 @@ public class LockFile {
         } finally {
             synchronized (LockFile.class) {
                 if (lock == null) {
-                    System.getProperties().remove(getVmLockKey());
+                    System.getProperties().remove(lockSystemPropertyName);
                 }
             }
         }
@@ -97,7 +111,7 @@ public class LockFile {
 
     /**
      */
-    public void unlock() {
+    synchronized public void unlock() {
         if (DISABLE_FILE_LOCK) {
             return;
         }
@@ -111,14 +125,17 @@ public class LockFile {
         if (lock != null) {
             try {
                 lock.release();
-                System.getProperties().remove(getVmLockKey());
             } catch (Throwable ignore) {
+            } finally {
+                if (lockSystemPropertyName != null) {
+                    System.getProperties().remove(lockSystemPropertyName);
+                }
+                lock = null;
             }
-            lock = null;
         }
         closeReadFile();
 
-        if (deleteOnUnlock) {
+        if (locked && deleteOnUnlock) {
             file.delete();
         }
     }
@@ -129,18 +146,42 @@ public class LockFile {
 
     private void closeReadFile() {
         // close the file.
-        if (readFile != null) {
+        if (randomAccessLockFile != null) {
             try {
-                readFile.close();
+                randomAccessLockFile.close();
             } catch (Throwable ignore) {
             }
-            readFile = null;
+            randomAccessLockFile = null;
+        }
+    }
+
+    /**
+     * @return true if the lock file's last modified does not match the locally cached lastModified, false otherwise
+     */
+    private boolean hasBeenModified() {
+        boolean modified = false;
+
+        //Create a new instance of the File object so we can get the most up to date information on the file.
+        File localFile = new File(file.getAbsolutePath());
+
+        if (localFile.exists()) {
+            if(localFile.lastModified() != lastModified) {
+                LOG.info("Lock file " + file.getAbsolutePath() + ", locked at " + new Date(lastModified) + ", has been modified at " + new Date(localFile.lastModified()));
+                modified = true;
+            }
+        }
+        else {
+            //The lock file is missing
+            LOG.info("Lock file " + file.getAbsolutePath() + ", does not exist");
+            modified = true;
         }
 
+        return modified;
     }
 
     public boolean keepAlive() {
-        return lock != null && lock.isValid() && file.exists();
+        locked = locked && lock != null && lock.isValid() && !hasBeenModified();
+        return locked;
     }
 
 }
